+++ membrane_curvature/__init__.py
@@ -4,7 +4,8 @@ MDAkit for Membrane Curvature
 """
 
 # Add imports here
-from .core import *
+from membrane_curvature.surface import normalized_grid, derive_surface, get_z_surface
+from membrane_curvature.curvature import mean_curvature, gaussian_curvature
 
 # Handle versioneer
 from ._version import get_versions
