+++ membrane_curvature/curvature.py
@@ -1,0 +1,53 @@
+import numpy as np
+
+def gaussian_curvature(Z):
+    """
+    Calculate gaussian curvature from Z cloud points.
+
+
+    Parameters
+    ----------
+    Z : Numpy array, cloud of points.
+
+
+    Returns
+    -------
+    K : 2d-array
+        Returns 2-dimensional array object with values of mean curvature.
+
+    """
+
+    Zy, Zx = np.gradient(Z)
+    Zxy, Zxx = np.gradient(Zx)
+    Zyy, _ = np.gradient(Zy)
+
+    K = (Zxx * Zyy - (Zxy ** 2)) / (1 + (Zx ** 2) + (Zy ** 2)) ** 2
+
+    return K
+
+
+def mean_curvature(Z):
+    """
+    Calculates mean curvature from Z cloud points.
+
+
+    Parameters
+    ----------
+    Z : Numpy array, cloud of points.
+
+
+    Returns
+    -------
+    H : 2d-array
+        Returns 2-dimensional array object with values of gaussian curvature.
+
+    """
+
+    Zy, Zx = np.gradient(Z)
+    Zxy, Zxx = np.gradient(Zx)
+    Zyy, _ = np.gradient(Zy)
+
+    H = (Zx**2 + 1) * Zyy - 2 * Zx * Zy * Zxy + (Zy**2 + 1) * Zxx
+    H = -H / (2 * (Zx**2 + Zy**2 + 1)**(1.5))
+
+    return H
