+++ membrane_curvature/surface.py
@@ -107,54 +107,4 @@ def normalized_grid(grid_z_coordinates, grid_norm_unit):
     return z_normalized
 
 
-def gaussian_curvature(Z):
-    """
-    Calculate gaussian curvature from Z cloud points.
 
-
-    Parameters
-    ----------
-    Z : Numpy array, cloud of points.
-
-
-    Returns
-    -------
-    K : 2d-array
-        Returns 2-dimensional array object with values of mean curvature.
-
-    """
-
-    Zy, Zx = np.gradient(Z)
-    Zxy, Zxx = np.gradient(Zx)
-    Zyy, _ = np.gradient(Zy)
-
-    K = (Zxx * Zyy - (Zxy ** 2)) / (1 + (Zx ** 2) + (Zy ** 2)) ** 2
-
-    return K
-
-
-def mean_curvature(Z):
-    """
-    Calculates mean curvature from Z cloud points.
-
-
-    Parameters
-    ----------
-    Z : Numpy array, cloud of points.
-
-
-    Returns
-    -------
-    H : 2d-array
-        Returns 2-dimensional array object with values of gaussian curvature.
-
-    """
-
-    Zy, Zx = np.gradient(Z)
-    Zxy, Zxx = np.gradient(Zx)
-    Zyy, _ = np.gradient(Zy)
-
-    H = (Zx**2 + 1) * Zyy - 2 * Zx * Zy * Zxy + (Zy**2 + 1) * Zxx
-    H = -H / (2 * (Zx**2 + Zy**2 + 1)**(1.5))
-
-    return H
